+++ mobile/lib/db/enum/conflict_algo.dart
